Added NPC overlay to screen
NPCs now appear merged in with the tile obs to give a more screen
like view as the player would have it.

NPCs that move are currently not added, though there is a flag
in _get_sprites that could be toggled to add them back. Currently
it seems to be slowing down training.