Fix reversed stat/swap battle/start menus
These two menu's are the same thing but are in different orders
depending on if you access them from the start menu or from
within a battle.

This fix address's that by checking for menu state during battle,
ie. we known which of the two menu states the menu is in and
overwrites the menu state text if the menu's match.