TEST: nn_3_chan_fcl_screen_pos_1_weight (TRY AGAIN!)
Extra FCL with screen and position matrix before concating them.

FCL for gamestate and action, then just screen and gamestate for
final concat.

NOTE: This was the best run but overfitted slightly at the end.
Its worth referencing in the future if there is a way to cut
out overfitting after 40-50M.