Add support for filtering menu selections
Menu's can be ignored via game.allow_menu_selection(), which
returns a bool if the pyboy cmd (Press A) is sent. Though it is
built to take other pyboy cmds for filtering should that be
needed in the future.

Needs API to call process_game_states() to keep internal state
updated to verify if in menu/which menu for filtering.