Add battle checkpoints for each starter
This is a rich set of save points from pallet to veridian forest.

Each starter is pair with some different caught wild mon, all are
left as low level as possible. There are also some different
item arrangements.

This set is designed to train battles from the start of the
game.