Fix battle reward on animation
Depending on ram refresh timing it was possible to get into a
state where it looks like the battle was won, b/c of 0 enemy
pokemon left, during the intro battle anim.

Checks for battle type flag as a condition to battle won, as
you must be fighting a known battle type to win and recieve
reward.

Also tweaked debug message for action from int to string for
easer action debugging.