Fix battle text from menu state flicker
The single value for ID'ing battle text wasn't enough and it'd
allow other state's to bleed through when battle text was on the
screen.

 Very tricky to figure this one out, there is no sole clear ID for
 battle text but we can infer it from a combo of other reg's.
Battle text pause get's it 50% of the time but there is a delay
sometimes which give false positive on ID'ing menu's. Text box id
work's the rest of the time but it shares a common key with
pokemon menu so we also can't trust that. The UNKNOWN_D730 reg
in battle is always 0x40 when in the pokemon menu though so it
let's us rule out the pokemon menu.

Days and days spent on figuring this out, if you can find a better
set of doc'd reg's then it should be updated but for now this is
the best within reason.