Add HM menu support
Can read HM's from pokemon submenu, this is REALLY hacky but
enables a much needed feature.

In the future it'd be good to rewrite the whole menu class to be
a dict of menu pointers keys containing val which is also a dict
of known states that share the menu pointer and that dicts val's
are 3/4 screen peaks to the right of the pointer as UIDs.

This would allow all menu lookups in 4/5 ram reads and scale very
well for any new menu as it's just adding the tile id's into the
screen peak dict.