Add trainer battle decay
Early in training battling trainers is super valuable for
learning how to fight. Once the AI is more grown up fighting
gets in the way of reaching new places fast.

This add's a large reward for winning training battles but then
decays winning fights with the reset count or 'age' obs.

This will need future turning but to start is achieving it's
goal of having the AI start to skip trainers on the way to
Mt. Moon.